fix(inspector_panel): fixed value's randomly flooring

+++ include/craftr_engine/vector2.hpp
@@ -10,7 +10,7 @@ public:
   Vector2 operator+(const Vector2 &other) const;
   Vector2 operator-(const Vector2 &other) const;
 
-  Vector2(int x, int y);
+  Vector2(float x, float y);
   Vector2();
 };
