+++ src/craftr_engine/vector2.cpp
@@ -1,6 +1,6 @@
 #include "craftr_engine/vector2.hpp"
 
-Vector2::Vector2(int x, int y) {
+Vector2::Vector2(float x, float y) {
   this->x = x;
   this->y = y;
 }
